+++ src/core/application.ts
@@ -12,6 +12,7 @@ import { Container } from "./container.js";
 import { Config, type ConfigData } from "./config.js";
 import { Router } from "./http/router.js";
 import { View } from "./view.js";
+import { Events } from "./events.js";
 import { ServiceProvider, type ProviderClass } from "./provider.js";
 import { setApplication } from "./helpers.js";
 
@@ -37,6 +38,7 @@ export class Application extends Container {
     this.singleton(Config, () => new Config());
     this.singleton(Router, (app) => new Router(app));
     this.singleton(View, () => new View());
+    this.singleton(Events, () => new Events());
   }
 
   path(...segments: string[]): string {